update PR issues
