+++ pyproject.toml
@@ -74,6 +74,11 @@ indent-style = "space"
 skip-magic-trailing-comma = false
 line-ending = "auto"
 
+[dependency-groups]
+dev = [
+    "ruff>=0.11.5",
+]
+
 [tool.ruff.lint.isort]
 known-first-party = ["garth"]
 combine-as-imports = true
