+++ src/garth/data/body_battery.py
@@ -1,304 +1,16 @@
-from __future__ import annotations
+"""Body Battery data module - imports from separate files."""
 
-from datetime import date, datetime, timedelta
-from functools import cached_property
-from typing import Any, List
+from .body_battery_events import BodyBatteryData, BodyBatteryEvent
+from .daily_body_battery_stress import (
+    BodyBatteryReading,
+    DailyBodyBatteryStress,
+    StressReading,
+)
 
-from pydantic.dataclasses import dataclass
-from typing_extensions import Self
-
-from .. import http
-from ..utils import format_end_date
-
-
-@dataclass
-class BodyBatteryEvent:
-    """Body Battery event data."""
-    event_type: str
-    event_start_time_gmt: datetime
-    timezone_offset: int
-    duration_in_milliseconds: int
-    body_battery_impact: int
-    feedback_type: str
-    short_feedback: str
-
-
-@dataclass
-class BodyBatteryReading:
-    """Individual Body Battery reading."""
-    timestamp: int
-    status: str
-    level: int
-    version: float
-
-
-@dataclass
-class StressReading:
-    """Individual stress reading."""
-    timestamp: int
-    stress_level: int
-
-
-@dataclass
-class DailyBodyBatteryStress:
-    """Complete daily Body Battery and stress data."""
-    user_profile_pk: int
-    calendar_date: str
-    start_timestamp_gmt: datetime
-    end_timestamp_gmt: datetime
-    start_timestamp_local: datetime
-    end_timestamp_local: datetime
-    max_stress_level: int
-    avg_stress_level: int
-    stress_chart_value_offset: int
-    stress_chart_y_axis_origin: int
-    stress_values_array: List[List[int]]
-    body_battery_values_array: List[List[Any]]
-
-    @cached_property
-    def body_battery_readings(self) -> List[BodyBatteryReading]:
-        """Convert body battery values array to structured readings."""
-        if not self.body_battery_values_array:
-            return []
-
-        readings = []
-        for values in self.body_battery_values_array:
-            if len(values) >= 4:
-                readings.append(BodyBatteryReading(
-                    timestamp=values[0],
-                    status=values[1],
-                    level=values[2],
-                    version=values[3]
-                ))
-        return readings
-
-    @property
-    def stress_readings(self) -> List[StressReading]:
-        """Convert stress values array to structured readings."""
-        if not self.stress_values_array:
-            return []
-
-        readings = []
-        for values in self.stress_values_array:
-            if len(values) >= 2:
-                readings.append(StressReading(
-                    timestamp=values[0],
-                    stress_level=values[1]
-                ))
-        return readings
-
-    @property
-    def current_body_battery(self) -> int | None:
-        """Get the latest Body Battery level."""
-        readings = self.body_battery_readings
-        return readings[-1].level if readings else None
-
-    @property
-    def max_body_battery(self) -> int | None:
-        """Get the maximum Body Battery level for the day."""
-        readings = self.body_battery_readings
-        return max(reading.level for reading in readings) if readings else None
-
-    @property
-    def min_body_battery(self) -> int | None:
-        """Get the minimum Body Battery level for the day."""
-        readings = self.body_battery_readings
-        return min(reading.level for reading in readings) if readings else None
-
-    @property
-    def body_battery_change(self) -> int | None:
-        """Calculate the Body Battery change for the day."""
-        readings = self.body_battery_readings
-        if not readings or len(readings) < 2:
-            return None
-        return readings[-1].level - readings[0].level
-
-    @classmethod
-    def get(
-        cls,
-        date_str: str | date | None = None,
-        *,
-        client: http.Client | None = None,
-    ) -> Self | None:
-        """Get complete Body Battery and stress data for a specific date."""
-        client = client or http.client
-        date_str = format_end_date(date_str)
-
-        path = f"/wellness-service/wellness/dailyStress/{date_str}"
-        response = client.connectapi(path)
-
-        if not isinstance(response, dict):
-            return None
-
-        # Parse timestamps
-        start_gmt = datetime.fromisoformat(
-            response.get("startTimestampGMT", "").replace("Z", "+00:00")
-        )
-        end_gmt = datetime.fromisoformat(
-            response.get("endTimestampGMT", "").replace("Z", "+00:00")
-        )
-        start_local = datetime.fromisoformat(
-            response.get("startTimestampLocal", "").replace("Z", "+00:00")
-        )
-        end_local = datetime.fromisoformat(
-            response.get("endTimestampLocal", "").replace("Z", "+00:00")
-        )
-
-        return cls(
-            user_profile_pk=response.get("userProfilePK", 0),
-            calendar_date=response.get("calendarDate", ""),
-            start_timestamp_gmt=start_gmt,
-            end_timestamp_gmt=end_gmt,
-            start_timestamp_local=start_local,
-            end_timestamp_local=end_local,
-            max_stress_level=response.get("maxStressLevel", 0),
-            avg_stress_level=response.get("avgStressLevel", 0),
-            stress_chart_value_offset=response.get(
-                "stressChartValueOffset", 0
-            ),
-            stress_chart_y_axis_origin=response.get(
-                "stressChartYAxisOrigin", 0
-            ),
-            stress_values_array=response.get(
-                "stressValuesArray", []
-            ),
-            body_battery_values_array=response.get(
-                "bodyBatteryValuesArray", []
-            )
-        )
-
-    @classmethod
-    def list(
-        cls,
-        end: date | str | None = None,
-        period: int = 1,
-        *,
-        client: http.Client | None = None,
-    ) -> List[Self]:
-        """Get Body Battery and stress data for multiple days."""
-        end_date = format_end_date(end)
-        all_data = []
-
-        for i in range(period):
-            date_to_fetch = end_date - timedelta(days=i)
-            data = cls.get(date_to_fetch, client=client)
-            if data:
-                all_data.append(data)
-
-        return all_data
-
-
-@dataclass
-class BodyBatteryData:
-    """Legacy Body Battery events data (sleep events only)."""
-    event: BodyBatteryEvent | None = None
-    activity_name: str | None = None
-    activity_type: str | None = None
-    activity_id: str | None = None
-    average_stress: float | None = None
-    stress_values_array: List[List[int]] | None = None
-    body_battery_values_array: List[List[Any]] | None = None
-
-    @property
-    def body_battery_readings(self) -> List[BodyBatteryReading]:
-        """Convert body battery values array to structured readings."""
-        if not self.body_battery_values_array:
-            return []
-
-        readings = []
-        for values in self.body_battery_values_array:
-            if len(values) >= 4:
-                readings.append(BodyBatteryReading(
-                    timestamp=values[0],
-                    status=values[1],
-                    level=values[2],
-                    version=values[3]
-                ))
-        return readings
-
-    @property
-    def current_level(self) -> int | None:
-        """Get the latest Body Battery level."""
-        readings = self.body_battery_readings
-        return readings[-1].level if readings else None
-
-    @property
-    def max_level(self) -> int | None:
-        """Get the maximum Body Battery level for the day."""
-        readings = self.body_battery_readings
-        return max(reading.level for reading in readings) if readings else None
-
-    @property
-    def min_level(self) -> int | None:
-        """Get the minimum Body Battery level for the day."""
-        readings = self.body_battery_readings
-        return min(reading.level for reading in readings) if readings else None
-
-    @classmethod
-    def get(
-        cls,
-        date_str: str | date | None = None,
-        *,
-        client: http.Client | None = None,
-    ) -> List[Self]:
-        """Get Body Battery events for a specific date."""
-        client = client or http.client
-        date_str = format_end_date(date_str)
-
-        path = f"/wellness-service/wellness/bodyBattery/events/{date_str}"
-        response = client.connectapi(path)
-
-        if not isinstance(response, list):
-            return []
-
-        events = []
-        for item in response:
-            # Parse event data
-            event_data = item.get("event")
-            event = None
-            if event_data:
-                event = BodyBatteryEvent(
-                    event_type=event_data.get("eventType", ""),
-                    event_start_time_gmt=datetime.fromisoformat(
-                        event_data.get("eventStartTimeGmt", "")
-                        .replace("Z", "+00:00")
-                    ),
-                    timezone_offset=event_data.get("timezoneOffset", 0),
-                    duration_in_milliseconds=event_data.get(
-                        "durationInMilliseconds", 0
-                    ),
-                    body_battery_impact=event_data.get("bodyBatteryImpact", 0),
-                    feedback_type=event_data.get("feedbackType", ""),
-                    short_feedback=event_data.get("shortFeedback", "")
-                )
-
-            events.append(cls(
-                event=event,
-                activity_name=item.get("activityName"),
-                activity_type=item.get("activityType"),
-                activity_id=item.get("activityId"),
-                average_stress=item.get("averageStress"),
-                stress_values_array=item.get("stressValuesArray"),
-                body_battery_values_array=item.get("bodyBatteryValuesArray")
-            ))
-
-        return events
-
-    @classmethod
-    def list(
-        cls,
-        end: date | str | None = None,
-        period: int = 1,
-        *,
-        client: http.Client | None = None,
-    ) -> List[Self]:
-        """Get Body Battery data for multiple days."""
-        end_date = format_end_date(end)
-        all_events = []
-
-        for i in range(period):
-            date_to_fetch = end_date - timedelta(days=i)
-            events = cls.get(date_to_fetch, client=client)
-            all_events.extend(events)
-
-        return all_events
+__all__ = [
+    "BodyBatteryData",
+    "BodyBatteryEvent", 
+    "BodyBatteryReading",
+    "DailyBodyBatteryStress",
+    "StressReading",
+]
